+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- The `go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc/filters` package to provide the way to filter the traces automatically generated in interceptors. (#2572)
+
 ## [1.8.0/0.33.0] - 2022-07-08
 
 ### Added